replica: Quote database fields
The DDL SQL of the replica managers quotes a number of fields that may 
be considered reserved in some databases. The replica manager however 
fails to quote the same fields in the SQL it generates at runtime. For 
postgresql this does not make a difference, however for other databases 
it results in errors caused by capitalization of field names.

Release notes: In a section of replica manager you should have a generic 
comment about compatibility with databases other than postgresql being 
improved.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4503/