billing, httpd: Fix creation of plots directory
For some reason the default for the path to the plots directory was
hard-coded. For this reason the tar package would still attempt to write
the plots to the system wide /var directory rather than the var
directory of dCache.

The plot directory was also missing from the installation. Instead the
httpd service batch file contains a lot of code explaining the admin how
to create the plot directory. I don't know why we went to such length
generating a log message rather than including the plots directory in
the default directory structure, but I am partly to be blamed for that
as I acked the original patch.

This patch fixes both issues.

Release notes: A directory for billing plots is now automatically created
during installation. For the tarball release the plots directory now
defaults to var/plots inside the installation directory.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4509/