WebAdmin: remove authentication from the Space Tokens page
Currently to access the Space Tokens page one was asked to login as
guest (or as admin). Since the guest username and password was written
as login message, there is no reason to have the guest login for this
page.
This patch removes the authentication from the Space Tokens page and
changes the login message.
(The guest login is kept in the code for further usage in the future in
case there will be a use case that requires a login weaker as admin and
stronger than everyone (not logged)).

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4299