WebAdmin: adding filters into CellInfo, Pool Queue and Active Transfers tables
This change allows users to filter the table content either by some
criteria typed in the "Quick Find"-field or by strings and numbers in
columns.
The filter supports AND, OR, NOT expressions, as well as Quotes, round
brackets for strings and "<","=", ">" and "!" signs for numbers.

The jQuery plugin called "The PicNet Table Filter" that is an open
source project released under The MIT License (MIT) is used. 
The current script had a bug.
It couldn't cope with tables with colspans in its header. The script
has been modified and the fix has been provided to the PicNet team.

RELEASE NOTES: filter content of CellInfo, Pool Queue and Active
Transfers tables.


Target: trunk
Request: 2.2
Acked-by: Tigran
patch: http://rb.dcache.org/r/4291/