replica: Drop postgresql specific code
Replica manager used an inherited table called 'shadow'. It appears to
be unused. One of the SQL scripts contains a comment saying it was used
to trick the planner.

If this code can be dropped then replica manager will only issue SQL
compliant statement and can be made to work with other databases.

Release notes: Add a generic comment about improving compatibility 
with alternative databases. Also describe that the shadow table is
no longer used or supported and that sites are encouraged to drop
the table from the database. Provide details how to do so.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4511/