webdav: Disable persistent connection between door and pool
This patch essentially fixes the same problem previously described in
r14921.

The problem is that a pool keeps a mover open for further requests until
the TCP connection is closed. Until that happens the
DoorTransferFinished message is not sent. The door uses
HttpUrlConnection when proxying HTTP GET requests and that class
internally caches connections to the server.

The previous commit attempted to solve the problem by calling the
disconnect method on the HttpUrlConnection object. Disconnect however
doesn't disconnect right away: In my tests it waits between 5 and 10
seconds (presumably to determine if the connection is really idle).

This patch adds a Connection: close header, disabling the use of a
persistent connection. The server will close the connection after the
file is sent.

Release notes: Fixes a problem in the WebDAV door in which proxied reads
would pause 5 to 10 seconds after the transfer.

Target: trunk
Request: 2.2
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4536/