webdav: Send correct reponse on authorization error
When attempting to perform a write operation that happens to be
disallowed for anonymous users, dCache would respond with '405 method
not allowed'. The correct response is '401 unauthorized'. A 405 response
is only appropriate if the door is marked read only and providing
credentials doesn't change the response.

Since 1.9.12 does not provide basic authentication, the patch is less
relevant for that branch and I propose not to merge it to 1.9.12.

Release notes: Fixed an authorization problem in the WebDAV door
relating to basic authentication. dCache errounously replied with
'405 method not allowed' when it should have replied '401 unauthorized'.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4551/