webdav: Fix compatibility with latest GVFS
The WebDAV PROPFIND method is used to query properties of resources and
collections. It allows a dynamic set of properties to be queried. RFC
2518 defines a set of standard properties, such as content length. A
PROPFIND request generates a multi status response. The properties in
the response are grouped into known properties and properties for which
an error is generated (either because they are unknown, undefined or
access is denied).

Milton places properties without a value (a null property) in the set of
known properties with an empty XML tag in the response. It is uncertain
whether this is allowed by RFC 2518, however recent versions of GVFS
(the Gnome virtual file system layer) reject such responses. It worked
in Ubuntu 10.04, but with the current Ubuntu 12.04 the WebDAV share
cannot be mounted in Gnome.

This patch works around the issue by preprocessing the response in a
custom response handler.

Release notes: Resolved a compatibility issue between the dCache WebDAV
door and Gnome. The problem prevented dCache WebDAV shares from being
mounted in Gnome.

Target: trunk
Request: 2.2
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4549/