Integrate JettyCell/Webadmin into legacy HttpServiceCell
There are currently two http/Jetty services, the legacy org.dcache.services.httpd.HttpServiceCell and org.dcache.admin.webadmin.jettycell.JettyCell.java.  The latter was added in connection with the administrative web application (webadmin).

This situation is undesirable from a number of standpoints, first and foremost because it violates DRY.  In addition:

(1) Webadmin is hardcoded to use JettyCell, both as a means to get a CellEndpoint and to extract configuration properties. This binds webadmin to JettyCell.

(2) JettyCell is also bound to webadmin in that it knows specifically about webadmin configuration properties.  

To resolve this problem, either the original HttpServiceCell needs to be provided with the capability of running servlets, or the legacy functionality provided by the HttpServiceCell needs to be integrated into the JettyCell, all the while ensuring that the service is decoupled from the application(s) it may run.

This patch adopts the first approach, eliminating JettyCell and turning the Jetty Web Application (servlet) into a new "alias" ["webapp"].

With this integration we have also altered the way properties and objects are accessed through JNDI:
  
   a.  All JNDI bindings provided during the construction of the web application alias (cell endpoint, nucleus, command-line arguments and options) are now scoped to the web application rather than being global to the domain in which the web application is running.

   b.  The way webadmin used to access JNDI was an ad hoc solution and unnecessary given the standard Spring support for extracting elements from JNDI directly.  This patch replaces the original mechanism with the Spring standard.

There are a couple of other issues with the current HttpServiceCell implementation. First, a single handler is given to the embedded Jetty server for processing all requests, and this handler then wraps each request in a fresh instance of a "service" whose responsibility it is to find the correct alias and execute its response handling. All of this, however, is written in a non-extensible way, with conditionals and specific auxiliary methods. This has been redesigned according to object-oriented principles of encapsulation. Moreover, there is really no utility in introducing another implicit API; all of this can be conceived of in terms of the (Jetty) handler API, with the delegating handler selecting the correct sub-handler and passing the task on to it. This also facilitates the integration of the new alias type based on org.eclipse.jetty.webapp.WebAppContext, which already implements this handler interface. 

One might wonder if these handlers may not actually be added directly to the Jetty server, since the setHandler() method provides for an array of such handlers. Because the server is constructed as part of the cell initialization called from its own constructor, however, these handlers are not yet available, and some modification of the cell semantics would be necessary to eliminate the intermediate handler; furthermore, an alias table of some sort must be preserved in order to service the cell's info(), ac_ls_alias_$_0_1() and ac_unset_alias_$_1() admin interface methods. (In the future we may wish to restructure HttpServiceCell to map 'alias' directly to Jetty's handlers and thus replace the custom request processing.)

Attached are two images illustrating the architectural changes.

The alternative http/https connectors implemented in JettyCell have been backported into HttpServiceCell so that one can use the secure admin logon.

RELEASE NOTES:

1.  It is no longer necessary to define a [webadminDomain] or a webadmin service.  The default webadmin.properties which are still relevant have been moved over to httpd.properties and the webadmin.batch file eliminated.  Note also that webadminAuthenticated is obsolete; if you do no wish any authentication (i.e., a read-only set of pages), set httpsPort to empty.

2.  The alias definition for webapp has this form:
     
    set alias <name> webapp <path-to-the-war-directory> <app-specific options ...>

3.  The webadmin page now is found at "/webadmin", because "/" continues to serve the static dCache Service page (they share the main httpdPort; default is 2288).

4.  Authentication to webadmin continues as before -- an administrative GID needs to be defined (default is 1000).  Username + password (e.g., kpwd) and PK12 certificate (uploaded into the browser) authentication should both work.  The default port for https is 8444.

Target: trunk@17292
Patch: http://rb.dcache.org/r/4518
Ack'd by:  Tigran