[gplazma] add guava-inspired static method
Guava includes several static methods that allow an assert-or-throw 
pattern in a single line: checkArgument, checkState, etc.  Using these 
methods cuts down on boiler-plate code, so improves readability.

There is a large number of places in gPlazma code where there is a 
similar pattern: a test is made and, if true, a 
AuthenticationException is thrown.  A Guava-inspired assert-or-throw 
static method would improve readability of the gPlazma code-base.

The names (both the class and the method) where chosen to reflect the 
Guava Preconditions class, so that a developer who is familiar with 
Guava should pick up what's happening easily.  These names may be 
suboptimal, but I haven't thought of any better names that have 
comparable length.

This patch also includes some minor refactoring and fixes some 
occurances of the log-and-throw anti-pattern.

Target: trunk
Patch: http://rb.dcache.org/r/4526
Acked-by: Gerd Behrmann