Page view unit tests for webadmin broken from changes in patch 4518.
The page Junit tests for webadmin all generate null pointer exceptions when calling the isAuthenticatedMode() method.

This is because the new method depends on the presence of the HttpServiceCell, which is not constructed for the purposes of these tests.

The fix is to override this method appropriately in the class providing this setup for the test harness.  Several other methods also needed to be overridden in the same manner.

target: trunk@17300
patch: http://rb.dcache.org/r/4570/
ack'd by: Tigran