plugins: Add experimental HSQLDB plugin
The patch adds a new top-level directory for dCache plugins. The
subprojects in that directory are not distributed with dCache. They are
rather packaged as plugins, installable in the dCache plugin directory.

The patch adds an experimental plugin for adding HSQLDB support. The
bulk of the plugin provides alternative liquibase changelogs and
overrides default properties of various services to let them use HSQLDB.
Once the plugin is installed, all services will default to HSQLDB.

Despite liquibase promising that this should not be necessary, I had to
make new changelogs specifically for HSQLDB for a number of databases:

- PinManager ran into a problem between DataNucleus, Liquibase, and
  HSQLDB. For HSQLDB, Liquibase turns all identifiers to upper case.
  PinManager is however configured to preserve the case as defind in the
  JDO mapping (which happens to be lower case). Thus there is no way
  I can make a regular changelog file for PinManager work with HSQLDB (not
  unless I change the case of the existing tables used with postgresql).

  I also ran into issues with the sequence and min value not being
  supported, but that's fixed in the latest HSQLDB.

  The workaround is to embed the SQL DDL in the changelog.

- For billing I ran into the problem that its changelog contains a
  precondition preventing it from running with any other DB than
  postgresql. This despite the basic schema working fine with other
  databases. Only the triggers are specific to postgresql. That situation
  should have been handled by marking specific changelogs for use with
  postgresql, not by a precondition for the complete changelog file.

  I added a new changelog without the triggers.

- For replica I added new changelogs. The existing SQL scripts for replica
  manager are very postgresql specific. The postgresql specific parts of
  the schema do however not appear to be used by replica manager.

  The changelogs should probably be updated to use Liquibase's XML for
  defining the schema. In their present form I simply embedded the
  existing SQL (with adjustments).

Unfortunately I cannot adjust the existing changelog files. This is
because liquibase stores checksums of the changelogs in the database.
Making any changes to already deployed changelogs will cause conflicts.

I consider the plugin experimental as I do not consider the changelog
files final. Any deployment using the plugin should expect that the
database has to be recreated when we update the changelog files.

The patch updates HSQLDB and LiquiBase to the latest versions. I had to
remove the setting of sql.enforce_strict_size to make the tests work 
with HSQLDB 2.2. The default of that property is true since version 2.0, 
so there is no reason to set it (I guess they removed the property 
in 2.2).

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4512/