pool: Decouple Netty components from mover implementations
The xrootd and http movers share a common architecture, both building on
Netty.

They both rely on a static Netty instance. The mover registers itself
with this static instance, and a Netty channel handler queries this
static instance using an UUID to obtain the mover reference. The channel
handler then uses open and close calls to mark the mover as being "in
use".

This architecture has several problems: Making it thread safe is
difficult as "opening" a mover is a two step process involving three
different components. Consequently both movers have fairly complicated
logic to elliminate all the potential race conditions. The channel
handler is also intimately linked to the mover implementation, making
unit testing more difficult and preventing the channel handler from
being reused.

This patch refactors the code in question.

The movers now register a MoverChannel with the static server instance.
A MoverChannel wraps the real RepositoryChannel, adding space accounting
(by delegating to an Allocator), tracking of transfer time and amount of
data transferred, and provides the protocol info and access mode of the
mover.

The channel handler extracts the MoverChannel from the static server and
performs all transfer operations using only the MoverChannel. The
channel handler has no knowledge about the mover.

The mover blocks on a call to the static server during transfer. Once
the channel handler releases the MoverChannel, the mover's thread is
waken up.

The patch also cleans a fair number of minor issues: Updates outdated
comments, moves fields and constants to the classes in which they are
used, etc.

The patch cuts some 250 lines of code, but more importantly it makes it
easier to add PUT support to the HTTP mover.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4530/