webdav,pool: Add content-location header
From RFC 2616: The Content-Location entity-header field MAY be
used to supply the resource location for the entity enclosed in
the message when that entity is accessible from a location
separate from the requested resource's URI.

This patchs adds such a content-location header to the reply for
HTTP GET requests sent to a pool.

The URI for the header is provided by the door as part of the
protocol info. For compatibility reasons and for pool to pool
transfers, this field is allowed to be null. In this case the
content-location header is not added.

It became more and more obvious that having the sendHTTPxxx
utility methods in the base class of the request handler is wrong:
We several times had to refine those utility methods to take
changes in the subclass into account. For that reason I have moved
those methods to the subclass.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4545