webdav: Make better use of Guava and avoid redundant transfer map
Guava has a convenient method to copy data between streams. This patch
makes use of it.

The WebDAV door maintained two maps with essentially the same
information. Previously we didn't make use of the session ID because
HttpProtocolInfo defines it as an int while it really is a long. However
the risk of collisions when wrapping is so low (and only exists if 
transfers last longer than 42 days) that it is worth simplifying the code.

In a future patch we should transition HttpProtocolInfo to use a long
value, although we have to consider backwards compatibility when doing
so.

Release notes: Not worth mentioning in the release notes.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4552/