webdav,pool: Add redirect on PUT
This patch adds support for redirecting HTTP PUT requests to the pool.
By doing so we avoid the need to proxy all data through the door. Only
clients that send an Expect: 100-Continue header will be redirected.

Milton has no support for redirecting PUT requests. We therefore
have to work around this limitation by throwing a redirect
exception and catch it in the filter invoking the Milton request
handlers.

Another workaround is needed for correct handling of the Expect:
100-continue handler in the door. Jetty sends the 100 continue
responds as soon as the servlet calls getInputStream. It
interprets this call as an indication that the servlet intents to
read the data. Milton however calls getInputStream before ever
calling our code and thus before we can generate the redirect
exception. The workaround is to make the call to getInputStream
lazy by only calling it once we invoke any of the methods on the
input stream.

These limitations are not bugs in Milton - Milton was not
designed for these usecases.

On the pool side the request handler had to be extended with
processing of PUT messages. To avoid excessive memory usage and/or
avoid limiting the upload size, the chunk aggregator had to be
removed from the channel pipeline. Hence we need to process
HttpChunks ourself in the request handler.

The maxChunkSize parameter is obsolete. It wasn't actually used to
control the max chunk size - it was used to limit the http message
size. Since we no longer use the aggregator, there is no need for
this parameter.

The HTTP mover now implements ChecksumMover. Digest computation is
done in ChecksumChannel, a wrapper for RepositoryChannel. It is
conceivable that this class can be useful in other movers too.

Release notes: Added option to redirect HTTP uploads. When enabled,
data is transfered directly from the client to the pool without
using the door as a proxy. Set webdav.redirect.on-write to false to
disable this feature.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/4553/