Add system-test package for building ready-to-use dCache
This patch adds a new system-test package. It builds a pre-configured
ready-to-use dCache that can be started directly in the build directory.
The installation will be put in modules/system-test/target/dcache. The
file layout is very similar to that of the tarball build. You can start
dCache by running modules/system-test/target/dcache/bin/dcache start.

To compile use

    mvn package -am -pl modules/system-test/ -DskipTests

To selectively clean the system-test build use

    mvn clean -pl modules/system-test/ -DskipTests

Care has been taken to prevent Maven from overwriting your modified
configuration files in the build directory - unless you modified the
corresponding source files such that they are newer than the file in the
build directory. Clean the build directory if you want to force your
changes to be overwritten.


dCache is configured as a single domain instance with as many services
configured as possible. It uses embedded HSQLDB for all databases.

The only file stored outside the build directory is the PID file. It is
stored in /tmp. This is to prevent that cleaning the build directory
deletes the PID file.

The host certificate and ssh keys are embedded in the package. The goal
is to have a system that developers can use for quick and easy testing.
Hence I consider reusing certificates acceptable. The certificate is
valid for one year and will have to be regenerated periodically. The
certificate is generated by a self-signed CA for which the key has
been thrown away. The certificate is valid for use with the host name
'localhost'.

Upon first use you have to copy the files in modules/system-test/src/main/
skel/etc/grid-security/certificates to /etc/grid-security/certificates.
Otherwise your client tools will reject the server.

Upon first login to the admin door you have to confirm the use of the
ssh server key.

Both the above steps only have to be done once (at least until a year
from now) - that's the benefit of not regenerating the certificates
and keys upon every build.

The build will copy your CA certificates from
/etc/grid-security/certificates into the dCache installation. If you
have many CA certificates then this step will be slow. Reduce the number
of CA certificates to speed up the process.

The build will detect your DN and configure gPlazma to allow you access
to your dCache. Since your CA certificates are imported, you can use
your normal grid certificate for authentication. The build will also use
your username and UID. Password authentication will be configured,
although using the fixed password 'password'.

Most databases will be initialized during the build by running the
'dcache database update' command.

Several pools are created and have been configured for different
purposes. pool_write is a write pool for both disk and tape files.
pool_read is a read pool for both disk and tape files. Both are attached
to a simulated HSM. You need ruby installed for the HSM simulation to
work. Pool manager has been configured to force a pool to pool transfer
for disk files between the write pool and read pool (thus allowing you
to test the pool to pool code). For HSM files, the read can only be
satisfied by staging the file from tape (allowing you to test the HSM
code). The pools pool_r1, pool_r2, and pool_r3 are configured for use
with replica manager. The pool pool_sm is configured for use with space
reservations. It is attached to the same simulated tape system. For
staging the regular read pool is used (outside of space management).

A name space structure is automatically set up. The directories disk,
tape, resilient, reserved, public and private are created. They use
storage groups bound to the various pools, and have corresponding
defaults for retention policy and access latency.

A link group containing pool_sm is configured. No reservation are
created by default. Implicit space reservation is not enabled.

Some pools use the berkeley db repository, others the control directory
based repository (allowing both to be tested).

All doors except the kerberized doors and NFS doors are enabled.
Kerberized doors require an external service and NFS doors can only run
as root. Webdav is enabled for both HTTP and HTTPS. All monitoring
services are enabled. You can use your grid-certificate in the browser
for authentication or use password based authentication. Doors are
configured to only listen on the loopback interface (which should make
the setup reasonably secure). A few doors run on non-standard ports -
their default port conflicts with other doors.

Some services don't respect the listen option (admin-ssh2, httpd,
webadmin, webdav, xrootd) and for other services it doesn't apply
(movers in pools). For the first list we have to write fixes.

Note that all services run in a single domain. Cell tunnels have been
disabled. This is on purpose as a single domain setup is much easier to
test and debug (only one process to attach the debugger to, only one
domain to make a thread dump of, only one domain to make a heap dump of,
fewer resources used, faster startup and shutdown, only one log file).

The layout file is structured such that services are started in an order
that allows fast initialization. Several timing related settings are
tuned for quick response times, eg to quickly flush files to tape, to
quickly retry failures in pool manager, etc.

Billing is enabled to log to a database, however don't expect the plots
to work. Despite billing using DataNucleus (even using the more complex
JDO API, which in contrast to JPA is not limited to relation databases),
it still relies on (non-standard) database triggers. This is a shame as
in particular the large volumes of data generated by billing would be
obvious to store in a non-relational database (even more so when seeing
that the schema of the billing db is non-relational). This should have
beeen trivial to do with DataNucleus. That being said, HSQLDB does have
extensive support for triggers - we just have to rewrite them
specifically for HSQLDB.

The JVM is configured to enable debugging and listens for a connection
from a debugger on port 2299 on the loopback interface.

Note that SRM will attempt to create indexes on computed fields. This is
not supported by HSQLDB and the log file will contain error messages.
These can be safely ignored.

Event and trace logging are enabled at debug log level by default.

No automated tests are defined. In its present form this patch simply
provides an easy to use test dCache. I imagine that we can use this as a
foundation for a suite of automated system tests, but no effort towards
such a suite has been made.

I have included a modified version of my test script that I usually run
before committing stuff. Run it with -h to see available options. Note
that it will not pass for all services with the default options - eg
xrootd and dcap require public write permissions and will only work if
you run it against the public directory. Similarly running it for the
tape directory will run into timing issues because files have to be
flushed and staged to be readable.

Release notes: There are no changes to dCache in this patch and thus
there is nothing to note in the release notes.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4520/