vorolemap: update unit-tests
This patch refactors the unit-tests for the voRoleMap plugin as 
preparatory work for adding the missing functionality compared to 
gPlazma-1; this loss of functionality has been citied by Tier-1 sites 
as a reason for not moving to gPlazma2.

The class now follows the BDD-style of unit-testing.  In fact, the 
refactoring has removed the dependency on external fixtures.  
However, the files containing the fixtures have not been removed as 
other unit-tests still depend on them.

The refactored tests should cover the same functionality as the 
previous tests with the exception of some unrealistic scenarios 
(e.g., testing with two DNs).  Some additional scenarios are now 
tested.

There are some points that suggest the current VoRoleMap plugin does 
not function correctly.  When found, these have been documented with 
FIXME or REVISIT comments; however this patch is strictly limited to 
refactoring the tests, so no changes are made to the corresponding 
plugin.

Since the tests are intended to verify a future patch (that adds the 
missing functionality), it would be advantageous if this patch was 
also back-ported to the supported branches.

Target: trunk
Patch: http://rb.dcache.org/r/4576/
Acked-by: Tigran Mkrtchyan
Acked-by: Karsten Schwank