dcache: Remove ThreadManager unit tests
The tests fail randomaly. Most were already marked to be ignored,
although the remaining once fail randomly too.

As far as I can see, the class under test is not testable this
way: It uses lots of static fields and there can only really
be a single instance of it. It is probably better to get rid
of ThreadManager instead of trying to fix the tests.

Release notes: Nothing worth mentioning.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4588/