Integrate JettyCell/Webadmin into legacy HttpServiceCell, revised.
The revised patch makes the following modifications:

1.  The unmodifiable (read-only) context map has been changed to mutable so that cells can update the context and the static page contexts displayed.
2.  Potential concurrency issues on main alias map eliminated by using ConcurrentMap.
3.  Defaults for https are now properly handled:  a new property, 'authenticated', controls whether the cell looks up https port and the PKI properties.  
4.  The old webadmin.properties and webadmin.batch files have been restored and marked for obsolescence.
5.  Some minor changes (suggested below and marked as resolved) are also incorporated.
6.  I have eliminated the Fermilab header from all the files which derive in some way from the original HttpdServiceCell.

A second patch will follow in which this entire cell will be converted to a UniversalSpringCell (as perhaps it should have been from the outset), and in which some of the other suggestions (more use of Guava, for instance) will be incorporated.

I have tested this manually with the authenticated flag set to false and true.

I have tried to run the unit tests but have not been successful in figuring out how to get only the core and webadmin tests to run through ant (it always runs the gPlazma ones and hence the crash with the Argus tests).

RELEASE NOTES:

The user must now explicitly set the "authenticated" property to true for the httpd domain in order to get the authenticated web pages. This avoids bad behavior when the setup does not have a host certificate and https authentication is not desired. The static pages are still present with this patch (no redirection redirection to the webadmin servlet).  The patch also fixes a bug introduced by the original integration in which three of the static links would remain with the message "busy" and the page never produced because the context was not being updated.

Target: trunk
Patch: http://rb.dcache.org/r/4518
Acked-by: tigran