services: fix test for srmSpaceManagerEnabled
In 2.2, the "one-of" annotation was introduced which, for 
srmSpaceManagerEnabled, defines value values of 'true', 'on', 'yes' 
and 'enabled'.  Despite all these possible synonyms, the batch files 
for services that test the value of srmSpaceManagerEnabled check only 
'yes'.

Recent investigation from Dmitry showed that people are configuring 
srmSpaceManagerEnabled with 'true' rather than 'yes'.  (We might have 
encouraged this.)  With the srmSpaceManagerEnabled tests taking only 
'yes' to mean space-manager is enabled, misconfiguration is possible, 
with the space-manager running but not receiving all messages.

This patch fixes the test so that all the valid values give the same 
behaviour.  The patch also adjusts the SRM (so it has the same sense 
as other tests) and with transfermanager the patch removes the 
duplicate query.

Target: trunk
Request: 2.2
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=7171
Patch: http://rb.dcache.org/r/4582/
Acked-by: Dmitry Litvintsev