Billing package.jdo default VARCHAR specification inadequate for several table columns
RT 7207:  alter table while dcache is running

Problem was reported regarding long paths (> 256 chars) in the billing doorinfo table.  An attempt was made to alter the table's datatype manually, but dCache continued to enforce the 256 char limit.

The issue is that DataNucleus falls back to preset default lengths (in this case 256) for datatypes when the package.jdo does not specify this column attribute.

The bug fix involves the necessary change to the billing data package.jdo as well as the addition of a changeset to billing's liquibase xml.

Along with the doorinfo path column, the patch also modifies:

pnfsid               from 256 to 36
errormessage    from 256 to 1024

on all five of the main data tables.

Target:  trunk@17366
Patch: http://rb.dcache.org/r/4593/
Acked-by: Dmitry
Bug:  http://rt.dcache.org/Ticket/Display.html?id=7207&results=e56e3ea315fd9b2b5b53790ec737966a
Request: 1.9.12, 2.2

RELEASE NOTES:

The user should not have to do anything; however, if the installation is newer and was created by the original liquibase changelog, 
the changeset which alters the datatype lengths may take a while to complete on restart.  On test tables, which are of very modest size, the changes took 4 minutes.